+++ src/Repositories/Entities/Deadline.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Bot\Commands\Deadline;
+namespace Bot\Repositories\Entities;
 
 class Deadline
 {
